Fix background color being reset on select options
Due to the way the caret is rendered on the select box, the background
color is reset to default on select options when the select box loses
its :hover state. This is primarily noticeable on the dark theme.
Although I'm not sure what the best fix for this is, the simplest
solution is to just override the background color on option tags that
are children of non-:hover select boxes.