Fixed "Hide staff tools" not actually hiding all the staff tools (#619)
* Fixed option "Hide staff tools" not actually hiding staff tools

Code was expecting value to be a string, while it's actually being
parsed as boolean. Additionally, until settings are saved, value is
being parsed as empty string.

* Moved staff tools hiding logic from JS to CSS using new data attribute

After I've fixed the original intended staff tools hiding logic, I was
annoyed at how tools will "flash" on and off on every page load. So
to prevent that, I moved hiding logic from JS to simple CSS selector.
Server is creating special data attribute on the body element and CSS
looks for this attribute and hides anything marked by the
`js-staff-action` classname. JS logic is replaced with small "preview"
of this hiding when toggling the option in Local settings tab.

* Removed extension from store import

Co-authored-by: liamwhite <liamwhite@users.noreply.github.com>

---------

Co-authored-by: liamwhite <liamwhite@users.noreply.github.com>